separated the navBar
